+++ jpa/eclipselink.jpa.test/src/it/java/org/eclipse/persistence/testing/tests/jpa/xml/composite/advanced/EntityMappingsAdvancedJUnitTestCase.java
@@ -1732,7 +1732,7 @@ public class EntityMappingsAdvancedJUnitTestCase extends JUnitTestCase {
         pk.add(dealer.getId());
 
         DatabaseSessionImpl session = getDatabaseSession();
-        return (Integer) session.getDescriptor(Dealer.class).getOptimisticLockingPolicy().getWriteLockValue(dealer, pk, session);
+        return session.getDescriptor(Dealer.class).getOptimisticLockingPolicy().getWriteLockValue(dealer, pk, session);
     }
 
     protected List<Employee> createEmployeesWithUnidirectionalMappings(String lastName) {
